started subtraction and addition refactor

+++ CalculatorService.Core/Models/AddRequest.cs
@@ -8,6 +8,6 @@ namespace CalculatorService.Core.Models
 {
     public class AddRequest
     {
-        public IEnumerable<double> Addends { get; set; }
+        public IEnumerable<uint> Addends { get; set; }
     }
 }
